+++ R/helpers_MPN.R
@@ -1,9 +1,9 @@
 # Point estimate for MPN -------------------------------------------------------
 
-.ptEstimate <- function(positive, tubes, amount) {
+.ptEst_MPN <- function(positive, tubes, amount) {
   # https://www.fda.gov/Food/FoodScienceResearch/LaboratoryMethods/ucm109656.htm
   .scoreFnc <- function(lambda, positive, tubes, amount) {
-    # Score function. Find root to maximize likelihood.
+    #score function: Find root to maximize likelihood.
     LHS <- sum((positive * amount) / (1 - exp(-lambda * amount)))
     RHS <- sum(tubes * amount)
     LHS - RHS
@@ -24,13 +24,36 @@
   }
 }
 
+.ptEstAdj_MPN <- function(MPN, tubes, amount) {
+  #bias adjustment for MPN (Salama et al., 1978; Haas, 1989)
+  if (MPN == 0) {
+    return(0)
+  } else if (is.infinite(MPN)) {
+    return(NA)
+  } else {
+    amount2 <- amount ^ 2
+    amount3 <- amount ^ 3
+    lambda_V <- MPN * amount
+    one_or_more <- 1 - exp(-lambda_V)
+    zi  <- tubes * one_or_more
+    cosh_term <- cosh(lambda_V) - 1
+    D   <- sum((amount2 * zi) / (2 * cosh_term))
+    wi1 <- amount2 / (2 * (one_or_more ^ 2) * (D ^ 3))
+    wi2 <- (amount3 * zi * sinh(lambda_V)) / (cosh_term ^ 2)
+    wi2[is.nan(wi2)] <- 0  #Inf/Inf is NaN (denom goes to infinity more quickly)
+    wi2 <- sum(wi2)
+    wi3 <- amount3 / (one_or_more * cosh_term * (D ^ 2))
+    wi  <- wi1 * wi2 - wi3
+    return(MPN - 0.5 * sum(wi * zi * exp(-lambda_V)))
+  }
+}
 
 # Likelihood, LRT, etc. --------------------------------------------------------
 
 # See Ridout (1994) - "A Comparison of CI Methods..."
 # Section 2.2 "Intervals Based on the LRT"
 
-.logL <- function(lambda, positive, tubes, amount) {
+.logL_MPN <- function(lambda, positive, tubes, amount) {
   #log-likelihood
   lambda_amount <- lambda * amount
   first_term    <- positive * log(1 - exp(-lambda_amount))
@@ -38,29 +61,30 @@
   sum(first_term - second_term)
 }
 
-.logLR <- function(lambda, lambda_hat, positive, tubes, amount) {
+.logLR_MPN <- function(lambda, lambda_hat, positive, tubes, amount) {
   #log-likelihood ratio
-  logL_alt  <- .logL(lambda_hat, positive, tubes, amount)
-  logL_null <- .logL(lambda, positive, tubes, amount)
+  logL_alt  <- .logL_MPN(lambda_hat, positive, tubes, amount)
+  logL_null <- .logL_MPN(lambda, positive, tubes, amount)
   2 * (logL_alt - logL_null)
 }
 
-.logLRroot <- function(lambda, lambda_hat, positive, tubes, amount, alpha = .05) {
+.logLRroot_MPN <- function(lambda, lambda_hat, positive, tubes, amount,
+                           crit_val) {
   #Find roots to get LR confidence limits
-  log_like_ratio <- .logLR(lambda, lambda_hat, positive, tubes, amount)
-  crit_val       <- qchisq(alpha, df = 1, lower.tail = FALSE)
+  log_like_ratio <- .logLR_MPN(lambda, lambda_hat, positive, tubes, amount)
   log_like_ratio - crit_val
 }
 
 
 # Confidence intervals ---------------------------------------------------------
 
-.jarvisCI <- function(MPN, positive, tubes, amount, sig_level) {
+.jarvisCI_MPN <- function(MPN, positive, tubes, amount, conf_level) {
   # See Jarvis (2010) - "Reconsideration of the derivation of..."
   all_negative <- sum(positive) == 0
   all_positive <- identical(positive, tubes)
   var_logMPN <- NA
   variance   <- NA
+  sig_level  <- 1 - conf_level
 
   if (!all_negative && !all_positive) {
     exp_term <- exp(-amount * MPN)
@@ -91,23 +115,24 @@
   list(variance = variance, var_logMPN = var_logMPN, LB = LB, UB = UB)
 }
 
-.likeRatioCI <- function(MPN, positive, tubes, amount, sig_level) {
+.likeRatioCI_MPN <- function(MPN, positive, tubes, amount, conf_level) {
   #Likelihood ratio confidence limits
   #See Ridout (1994) - "A Comparison of CI Methods..."
   all_negative <- sum(positive) == 0
   all_positive <- identical(positive, tubes)
+  crit_val <- qchisq(conf_level, df = 1, lower.tail = TRUE)
   if (all_negative || all_positive) {
-    jarvis_bounds <- .jarvisCI(MPN, positive, tubes, amount, sig_level)
+    jarvis_bounds <- .jarvisCI_MPN(MPN, positive, tubes, amount, conf_level)
     LB <- jarvis_bounds$LB
     UB <- jarvis_bounds$UB
   } else {
-    LB <- uniroot(.logLRroot, interval = c(1e-04, MPN),
+    LB <- uniroot(.logLRroot_MPN, interval = c(1e-04, MPN),
                   lambda_hat = MPN, positive = positive, tubes = tubes,
-                  amount = amount, alpha = sig_level,
+                  amount = amount, crit_val = crit_val,
                   extendInt  = "no", maxiter = 1e+04)$root
-    UB <- uniroot(.logLRroot, interval = c(MPN, 5 * MPN),
+    UB <- uniroot(.logLRroot_MPN, interval = c(MPN, 5 * MPN),
                   lambda_hat = MPN, positive = positive, tubes = tubes,
-                  amount = amount, alpha = sig_level,
+                  amount = amount, crit_val = crit_val,
                   extendInt  = "upX", maxiter = 1e+04)$root
   }
   list(LB = LB, UB = UB)
@@ -116,7 +141,7 @@
 
 # Blodgett's rarity index ------------------------------------------------------
 
-.rarity <- function(MPN, positive, tubes, amount) {
+.rarity_MPN <- function(MPN, positive, tubes, amount) {
 
   all_negative <- sum(positive) == 0
   all_positive <- identical(positive, tubes)